backend/requirements/base.txt
Django==5.2.1
djangorestframework==3.16.0
psycopg2==2.9.10
django-filter==25.1
django-environ==0.12.0
django-cors-headers==4.7.0